+++ src/players/generic/managers/ConfigurationManager.ts
@@ -1,5 +1,7 @@
 import {JWPlayerConfiguration} from "../../jwplayer/JWPlayerConfiguration";
+import {ConsentManagerConfig} from "./ConsentManager/ConsentManagerConfig";
 
 export class ConfigurationManager {
     public readonly jwPlayer = new JWPlayerConfiguration();
+    public readonly consent = new ConsentManagerConfig();
 }